put all atom-vllm related metadata builders into metadata.py
